# backend/docs/swagger.yaml
definitions:
...
  dto.SaveSiteSettingRequest:
    properties:
      about_text:
        type: string
      logo:
        type: string
      site_name:
        type: string
      site_subtitle:
        type: string
      theme_color:
        type: string
    type: object
  dto.SaveTagRequest:
    properties:
      name:
        type: string
      slug:
        type: string
    required:
    - name
    - slug
    type: object
...
  response.Body:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /api/v1/admin/site/settings:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Body'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Body'
      summary: 获取站点设置
      tags:
      - Site
    put:
      consumes:
      - application/json
      parameters:
      - description: 站点设置
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/dto.SaveSiteSettingRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Body'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Body'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Body'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - BearerAuth: []
      summary: 更新站点设置
      tags:
      - Site
  /api/v1/admin/tags:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Body'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Body'
      summary: 获取标签列表
      tags:
      - Tag
    post:
      consumes:
      - application/json
      parameters:
      - description: 标签参数
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/dto.SaveTagRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Body'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Body'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Body'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - BearerAuth: []
      summary: 创建标签
      tags:
      - Tag
  /api/v1/admin/tags/{id}:
    delete:
      parameters:
      - description: 标签ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Body'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Body'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Body'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Body'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - BearerAuth: []
      summary: 删除标签
      tags:
      - Tag
    put:
      consumes:
      - application/json
      parameters:
      - description: 标签ID
        in: path
        name: id
        required: true
        type: integer
      - description: 标签参数
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/dto.SaveTagRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Body'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Body'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Body'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Body'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - BearerAuth: []
      summary: 更新标签
      tags:
      - Tag
  /api/v1/admin/upload/image:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: 图片文件
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Body'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Body'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Body'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - BearerAuth: []
      summary: 上传单图
      tags:
      - Upload
  /api/v1/admin/upload/images:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: 图片文件(可多选)
        in: formData
        name: files
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Body'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Body'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Body'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - BearerAuth: []
      summary: 上传多图
      tags:
      - Upload
...
  /api/v1/site/settings:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Body'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Body'
      summary: 获取站点设置
      tags:
      - Site
  /api/v1/tags:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Body'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Body'
      summary: 获取标签列表
      tags:
      - Tag
...
swagger: "2.0"